python-voice-agent/requirements.txt
livekit-agents>=0.8.0
livekit-plugins-openai>=0.6.0
livekit-plugins-elevenlabs>=0.6.0
livekit-plugins-silero>=0.6.0
httpx>=0.27.0
python-dotenv>=1.0.0
